+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Disallow redirects on POST/DELETE/... requests and require status code 200 on `POST /result` requests. 
   This improves error information where `POST /result` would involve a redirect. (EP-3889)
+- Pre-resolve redirection of root URL in version discovery. (EP-3889)
 
 ### Removed